+++ src/test/java/subway/domain/station/SectionsTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.Test;
+import subway.domain.section.Section;
+import subway.domain.section.Sections;
 
 @SuppressWarnings("NonAsciiCharacters")
 class SectionsTest {
